fix: resolve ReflectionResult AttributeError and improve attribute access patterns
- Fixed incorrect .get() method usage on ReflectionResult objects in enhanced_researcher.py and nodes.py
- Changed dictionary-style access to proper attribute access for Pydantic BaseModel
- Updated reflection_insights access patterns to use hasattr() and getattr() where appropriate
- Applied code formatting with black to maintain consistency

+++ src/utils/performance/performance_optimizer.py
@@ -1137,7 +1137,7 @@ class AdvancedParallelExecutor:
 class AdaptiveRateLimiter:
     """Adaptive rate limiter that adjusts based on system load."""
 
-    def __init__(self, initial_rate: int = 100, time_window: int = 60):
+    def __init__(self, initial_rate: float = 100.0, time_window: int = 60):
         self.initial_rate = initial_rate
         self.current_rate = initial_rate
         self.time_window = time_window
